add more chunk & block editing doc

+++ crates/py_bindings/examples/main.py
@@ -4,15 +4,17 @@ from inventory import inventory
 from the_blockheads_tools_py import WorldDb, Arch
 import pathlib
 import shutil
+import sys
 
 world_db = WorldDb.open_path("../tests/resources/9c3/world_db")
 world_db_io(world_db)
 chunk_block(world_db)
-inventory(world_db)
+# inventory(world_db)
 
-folder_path = pathlib.Path("./output")
-if folder_path.exists() and folder_path.is_dir():
-    shutil.rmtree(folder_path)
-    print(f"Deleted existing folder: {folder_path}")
-folder_path.mkdir(parents=True, exist_ok=True)
-world_db.save_path(folder_path, Arch.Arch32)
+if len(sys.argv) > 1:
+    folder_path = pathlib.Path(sys.argv[1])
+    if folder_path.exists() and folder_path.is_dir():
+        shutil.rmtree(folder_path)
+        print(f"Deleted existing folder: {folder_path}")
+    folder_path.mkdir(parents=True, exist_ok=True)
+    world_db.save_path(folder_path, Arch.Arch32)
